requirements.txt
python-telegram-bot
nest-asyncio
